Assignment_2/Module_B/requirements.txt
Flask==2.3.3
PyJWT==2.12.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Werkzeug==2.3.7
